pyproject.toml
[project]
name = "financeapi"
version = "0.1.0"
description = "Personal finance API (MVP) - accounts, transactions, imports, categorization rules, insights"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "SQLAlchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9",
  "alembic>=1.13.2",
  "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py314"
